refactor(settings): extract GeneralPageController
Ninth slice. General page surface moves into GeneralPageController:
- Rendering-backend picker data (options, translated display names,
  startup snapshot)
- Animation-slider bounds (duration / min-distance / stagger-interval)

The backend display-name translation and startup snapshot initialisation
move from SettingsController's ctor body into GeneralPageController's
ctor so the data isn't initialised in two places. The corresponding
SettingsController members (m_renderingBackendDisplayNames,
m_startupRenderingBackend) are removed.

EasingSettings rebinds animation bounds through the sub-controller.

importAllSettings / exportAllSettings / daemonRunning intentionally stay
on SettingsController — those are global app actions, not General-page
state.
